requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
Pillow>=10.0.0
pandas>=2.0.0
opencv-python-headless>=4.8.0
scikit-learn>=1.3.0

dagster>=1.6.0
dagster-webserver>=1.6.0
google-cloud-storage>=2.14.0

fastapi>=0.109.0
uvicorn>=0.27.0
python-multipart>=0.0.6
celery>=5.3.0
redis>=5.0.0

pytest>=7.4.0
httpx>=0.26.0
